Removed distributed communication class
Added missing parameter to test

Fixed formatting

docs: fix build issues and add sub-sections (#69)
