+++ docs/axonn_style.py
@@ -1,0 +1,16 @@
+# Copyright 2024 Parallel Software and Systems Group, University of Maryland.
+# See the top-level LICENSE file for details.
+#
+# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+# The name of the Pygments (syntax highlighting) style to use.
+from pygments.styles.default import DefaultStyle
+from pygments.token import Generic
+
+
+# modifications to the default style
+class AxonnStyle(DefaultStyle):
+    styles = DefaultStyle.styles.copy()
+    background_color = "#f4f4f8"
+    styles[Generic.Output] = "#355"
+    styles[Generic.Prompt] = "bold #346ec9"
