+++ docs/examples.rst
@@ -1,0 +1,17 @@
+
+********
+Examples
+********
+
+Training
+============
+
+
+
+Fine-tuning
+===========
+
+
+
+Inference
+=========
