+++ docs/user_guide.rst
@@ -1,0 +1,28 @@
+**********
+User Guide
+**********
+
+Initializing AxoNN
+==================
+
+Tensor with Easy API
+====================
+
+Tensor using Advanced API
+=====================================
+
+Combining Tensor in AxoNN with PyTorch DDP
+==========================================
+
+Integration with other Parallel APIs
+====================================
+
+Huggingface
+-----------
+
+Pipelining in AxoNN 
+===================
+
+Combining Pipelining in AxoNN with Data Parallelism 
+===================================================
+
